% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult.R
\docType{data}
\name{adult}
\alias{adult}
\alias{get_adult_task}
\alias{adult_dataset}
\title{Adult Dataset}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/adult}
}
\usage{
get_adult_task()
}
\description{
Dataset used to predict whether income exceeds $50K/yr based on census data. Also known as "Census Income" dataset
Train dataset contains 13 features and 30178 observations.
Test dataset contains 13 features and 15315 observations.
Target column is "Target".
}
\section{Pre-processing}{

\itemize{
\item Country column has been removed since we expect them have no predictive power.
\item fnlwgt column has been removed since we expect them have no predictive power.
\item Rows contain NA in Workclass and Occupation has been removed.
}
}

\examples{
data("adult", package = "mlr3fairness")

}
\keyword{data}
