% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{binary_measure_score}
\alias{binary_measure_score}
\title{Binary Measure Score}
\usage{
binary_measure_score(prediction, base_measure, data_task)
}
\arguments{
\item{prediction}{The predictions made by the learner}

\item{base_measure}{The base measures}

\item{datatask}{The datatask for groupwise operations}
}
\value{
c(measure_score_1, measure_score_2)
}
\description{
This helper function returns a list of 2 measure scores based on a binary protected feature in datatask.
}
