% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_operations.R
\name{groupdiff_absdiff}
\alias{groupdiff_absdiff}
\title{Groupwise Absolute Differences}
\usage{
groupdiff_absdiff(x)
}
\arguments{
\item{x}{(´numeric()`)\cr
Measured performance in group 1, 2, ...}
}
\description{
Computes \eqn{max(abs(x-y, y-x))}, i.e. the smallest symmetric ratio between \eqn{x} and \eqn{y}
that is smaller than 1. If \eqn{x} is a vector, the symmetric ratio between all
elements in \eqn{x} is computed.
}
\examples{
groupdiff_absdiff(1:3)
}
