% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{groupdiff_tau}
\alias{groupdiff_tau}
\title{Groupwise Ratio}
\usage{
groupdiff_tau(x, y, eps = 0.01)
}
\arguments{
\item{x}{\link{´numeric`} \cr
Measured performance in group 1.}

\item{y}{\link{´numeric`} \cr
Measured performance in group 1.}

\item{eps}{\link{´numeric`} \cr
A small value to ensure no division by 0.}
}
\description{
Computes \code{min(x/y, y/x)} i.e. the symmetric ratio between x and y
that is smaller than 1.
}
