% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{groupwise_abs_diff}
\alias{groupwise_abs_diff}
\title{Groupwise Absolute Differenence (Fairness Metrics)}
\usage{
groupwise_abs_diff(prediction, base_measure, task, ...)
}
\arguments{
\item{prediction}{(\code{PredictionClassif()})\cr The predictions of the learner.}

\item{base_measure}{(\code{Measure()})\cr The base measures used to evaluate.}

\item{task}{(\code{TaskClassif()})\cr The data task for the fairness metric.}

\item{...}{Further arguments, currently ignored.}
}
\value{
Absolute difference of base measues between binary protected groups.
}
\description{
Groupwise Absolute Differenence (Fairness Metrics)
}
