% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupwise_diff.R
\name{groupwise_diff}
\alias{groupwise_diff}
\title{Groupwise Differenence (Fairness Metrics)}
\usage{
groupwise_diff(prediction, base_measure, data_task, ...)
}
\arguments{
\item{prediction}{(\code{PredictionClassif()})\cr The predictions of the learner.}

\item{base_measure}{(\code{Measure()})\cr The base measures used to evaluate.}

\item{data_task}{(\code{TaskClassif()})\cr The data task for the fairness metric.}

\item{...}{Further arguments, currently ignored.}
}
\value{
Difference of base measues between binary protected groups.
}
\description{
Groupwise Differenence (Fairness Metrics)
}
