% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupwise_quotient.R
\name{groupwise_quotient}
\alias{groupwise_quotient}
\title{Groupwise Quotient (Fairness Metrics)}
\usage{
groupwise_quotient(prediction, base_measure, data_task, ...)
}
\arguments{
\item{prediction}{The predictions of the learner.}

\item{base_measure}{The base measures used to evaluate.}

\item{data_task}{The data task for the fairness metric.}

\item{...}{Further arguments, currently ignored.}
}
\value{
Quotient of base measues between binary protected groups.
}
\description{
Groupwise Quotient (Fairness Metrics)
}
