% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_fairness_task}
\alias{is_fairness_task}
\title{Is this task a fairness task?}
\usage{
is_fairness_task(task)
}
\arguments{
\item{task}{\code{\link[mlr3:TaskClassif]{TaskClassif}}\cr
The data task used for evaluation}
}
\value{
result (\code{logical})
}
\description{
Is this task a fairness task?
}
\examples{
library(mlr3fairness)
library(mlr3)
task = tsk("compas")
is_fairness_task(task)
}
