% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureAvgIndividualFairness.R
\name{MeasureAvgIndividualFairness}
\alias{MeasureAvgIndividualFairness}
\title{Individual Fairness}
\description{
This measure \code{\link[mlr3:Measure]{mlr3::Measure()}} allows for computing summary individual fairness statistics.
See \link{individual_fairness} for computing individual level fairness scores.

Implements a fairness notion roughly requiring that "similar individuals should be treated similarly".
Mathematically, computing individual fairness requires defining two distance metrics \eqn{d_x}{dX} and \eqn{d_y}{dY}
defined on feature and prediction spaces respectively.
\deqn{
   d_y(f(x1), f(x2)) \leq d_x(x1, x2)
}{
   dY(f(x1), f(x2) <= dX(f1, x2))
}

This is computed by comparing differences between predictions and differences between features, be default using the
Gower distance. The user is encouraged to adapt the distance metric and features used for computation to form a
relevant distance metric that encodes ethical or regulatory considerations.
}
\details{
Both dX and dY need to be functions that take as input arguments \code{x}, \code{y} (two data.tables) and
an argument \code{cols} which specifies the colnames to be used for computing distances.
}
\examples{
# Create MeasureFairness to measure the Predictive Parity.
t = tsk("adult_train")
learner = lrn("classif.rpart", cp = .01)
learner$train(t)
measure = msr("fairness", base_measure = msr("classif.ppv"))
predictions = learner$predict(t)
predictions$score(measure, task = t)
}
\references{
Dwork, Cynthia, Hardt, Moritz, Pitassi, Toniann, Reingold, Omer, Zemel, Richard (2012).
\dQuote{Fairness through awareness.}
In \emph{Proceedings of the 3rd innovations in theoretical computer science conference}, 214--226.
Mukherjee, Debarghya, Yurochkin, Mikhail, Banerjee, Moulinath, Sun, Yuekai (2020).
\dQuote{Two Simple Ways to Learn Individual Fairness Metrics from Data.}
In \emph{International Conference on Machine Learning}, 7097--7107.
PMLR.
}
\section{Super class}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{MeasureAvgIndividualFairness}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dX}}{\code{\link{function}}
Distance function between observations.}

\item{\code{dY}}{\code{\link{function}}
Distance function between predictions.}

\item{\code{fraction}}{\code{\link{function}}
Fraction of points to use for distance computation.
Set to smaller values for larger datasets.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MeasureAvgIndividualFairness$new()}}
\item \href{#method-clone}{\code{MeasureAvgIndividualFairness$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate">}\href{../../mlr3/html/Measure.html#method-aggregate}{\code{mlr3::Measure$aggregate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format">}\href{../../mlr3/html/Measure.html#method-format}{\code{mlr3::Measure$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help">}\href{../../mlr3/html/Measure.html#method-help}{\code{mlr3::Measure$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print">}\href{../../mlr3/html/Measure.html#method-print}{\code{mlr3::Measure$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score">}\href{../../mlr3/html/Measure.html#method-score}{\code{mlr3::Measure$score()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureAvgIndividualFairness$new(
  id = "fairness.individual",
  task_type,
  predict_type,
  dX = NULL,
  dY = NULL,
  fraction = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character})\cr
The measure's id. Set to 'fairness.individual' if ommited.}

\item{\code{task_type}}{\link{`character} Task type to construct the metric for.}

\item{\code{predict_type}}{\link{`character``} Predict type to construct the metric for.}

\item{\code{dX}}{[\verb{function] distance metric for features. Defaults to }gower::gower_dist\verb{. See }details` for additional info.}

\item{\code{dY}}{[\verb{function] distance metric for predictions. Defaults to }gower::gower_dist\verb{.  See }details` for additional info.}

\item{\code{fraction}}{\link{`numeric} For large datasets: Should computation be approximated by computing on only a fraction? Initialized to 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureAvgIndividualFairness$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
