% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{measures}
\alias{measures}
\title{Measure Registry}
\format{
An object of class \code{environment} of length 3.
}
\usage{
measures
}
\description{
The \code{\link[=environment]{environment()}} \code{measures} keeps track of all measures in this package.
It stores meta information such as minimum, maximum or if the
measure must be minimized or maximized.
}
\examples{
names(measures)
measures$tpr
}
\keyword{datasets}
