% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSubgroup.R
\name{MeasureSubgroup}
\alias{MeasureSubgroup}
\title{Evaluate a metric on a subgroup}
\description{
Allows for calculation of arbitrary \code{\link[mlr3:Measure]{mlr3::Measure()}}s on a selected sub-group.
}
\examples{
\dontshow{if (rlang::is_installed("rpart")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library("mlr3")
# Create MeasureFairness to measure the Predictive Parity.
t = tsk("adult_train")
learner = lrn("classif.rpart", cp = .01)
learner$train(t)
measure = msr("subgroup", base_measure = msr("classif.acc"), subgroup = "Female")
predictions = learner$predict(t)
predictions$score(measure, task = t)
\dontshow{\}) # examplesIf}
}
\seealso{
\link{MeasureFairness}, \link{groupwise_metrics}
}
\section{Super class}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{MeasureSubgroup}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{base_measure}}{(\code{mlr3::Measure()})\cr
The base measure to be used by the fairness measures,
e.g. \code{\link[mlr3:mlr_sugar]{mlr3::msr()}} with "classif.fpr" for the false positive rate.}

\item{\code{subgroup}}{(\code{character})|(\code{integer})\cr
Subgroup identifier.}

\item{\code{intersect}}{(\code{logical})\cr
Should groups be intersected?}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasureSubgroup-new}{\code{MeasureSubgroup$new()}}
\item \href{#method-MeasureSubgroup-clone}{\code{MeasureSubgroup$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureSubgroup-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureSubgroup-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureSubgroup$new(id = NULL, base_measure, subgroup, intersect = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character})\cr
The measure's id. Set to 'fairness.<base_measure_id>' if ommited.}

\item{\code{base_measure}}{(\code{Measure()})\cr
The measure used to measure fairness.}

\item{\code{subgroup}}{(\code{character})|(\code{integer})\cr
Subgroup identifier. Either value for the protected attribute or position in \code{task$levels}.}

\item{\code{intersect}}{\code{\link{logical}} \cr
Should multiple pta groups be intersected? Defaults to \code{TRUE}.
Only relevant if more than one \code{pta} columns are provided.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureSubgroup-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureSubgroup-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureSubgroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
