% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3fairness-package}
\alias{mlr3fairness}
\alias{mlr3fairness-package}
\title{mlr3fairness: Fairness Package in mlr3 family}
\description{
This new package could be used by many other mlr3 subsystems like mlr3pipelines,
    mlr3tuning and mlr3proba. Then further help the researchers and industrial users to detect
    the fairness of datasets or algorithmic bias more easily, which could benefit the accuracy
    and impartiality of their models built by mlr3.
}
\author{
\strong{Maintainer}: Wei Siyi \email{weisiyi2@gmail.com}

}
