% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{mlr3fairness_assertions}
\alias{mlr3fairness_assertions}
\title{Assertion for mlr3fairness}
\description{
Functions intended to be used in packages extending \pkg{mlr3fairness}.
Most assertion functions ensure the right class attrbiture, and optionally additional properties.

If an assertion fails, an exception is raised.
Otherwise, the input object is returned invisibly.
}
\keyword{internal}
