% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_fairness.R
\docType{data}
\name{mlr_learners_fairness}
\alias{mlr_learners_fairness}
\title{Fair Learners in mlr3}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 5 rows and 3 columns.
}
\usage{
mlr_learners_fairness
}
\description{
Fair Learners in mlr3
}
\section{Predefined measures}{

\CRANpkg{mlr3fairness} comes with a set of predefined fairn learners listed below:

\tabular{lll}{
\strong{key} \tab \strong{package} \tab \strong{reference} \cr
regr.fairfrrm     \tab fairml \tab Komiyama et al., 2018\cr
classif.fairfgrrm \tab fairml \tab Komiyama et al., 2018\cr
regr.fairzlm      \tab fairml \tab Zafar et al., 2019   \cr
classif.fairzlrm  \tab fairml \tab Zafar et al., 2019   \cr
classif.fairnclm  \tab fairml \tab Scutari et al., 2021
}
}

\examples{
Available learners:
mlr_learners_fairness$key
}
\keyword{datasets}
