% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureFairness.R
\name{mlr_measures_fairness}
\alias{mlr_measures_fairness}
\alias{MeasureFairness}
\title{Fairness Measure}
\description{
This measure specializes \link{Measure} to allow for measuring statistical group fairness:
A common approach to quantifying a model's fairness is to compute the difference between a protected and an unprotected group according to some performance metric such as \verb{classification error}, \verb{false positive rate} or others.
This measure allows for the composition of a fairness-metric from such a \code{base_measure} by computing group-wise difference according to some strategy.
Several options wrt. computing difference such as \code{\link[=groupwise_abs_diff]{groupwise_abs_diff()}}, \code{\link[=groupwise_diff]{groupwise_diff()}}, or \code{\link[=groupwise_quotient]{groupwise_quotient()}} are available.

Predefined measures can be found in the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_measures}.
Predefined operations can be found in the \link{mlr3fairness}.
}
\section{Predefined measures}{

\code{mlr3fairness} comes with a set of predefined fairness measures listed below.
For full flexibility, \code{\link{MeasureFairness}} can be used to construct classical
group fairness measures based on a difference between a performance metrics across groups
by combining a performance measure with an operation for measuring differences.

A set of widely used metrics is included in mlr3fairness.
\itemize{
\item \code{fairness.fpr} :: Abs. difference in false positive rates across groups
\item \code{fairness.acc} :: Abs. difference in accuracy across groups
....
}
}

\examples{
# Create MeasureFairness to measure the Predictive Parity.
library(mlr3)
data_task = tsk("adult_train")
learner = lrn("classif.rpart", cp = .01)
learner$train(data_task)
measure = msr("fairness", base_measure = msr("classif.ppv"), operation = "groupwise_quotient")
predictions = learner$predict(data_task)
predictions$score(measure, task = data_task)
# Predfined measures:
msr("fairness.fpr")
}
\section{Super class}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{MeasureFairness}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{base_measure}}{(\code{Measure()})\cr
The base measure to be used by the fairness measures,
e.g. \link{mlr_measures_classif.fpr} for the false positive rate.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MeasureFairness$new()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate">}\href{../../mlr3/html/Measure.html#method-aggregate}{\code{mlr3::Measure$aggregate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format">}\href{../../mlr3/html/Measure.html#method-format}{\code{mlr3::Measure$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help">}\href{../../mlr3/html/Measure.html#method-help}{\code{mlr3::Measure$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print">}\href{../../mlr3/html/Measure.html#method-print}{\code{mlr3::Measure$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score">}\href{../../mlr3/html/Measure.html#method-score}{\code{mlr3::Measure$score()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureFairness$new(operation = "groupwise_abs_diff", base_measure)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{operation}}{(\code{character()})\cr
The operation name performed on the base measures.
\itemize{
\item Possible inputs should be one of:
\itemize{
\item "groupwise_abs_diff"
\item "groupwise_diff"
\item "groupwise_quotient"
}
}}

\item{\code{base_measure}}{(\code{Measure()})\cr
The measure used to perform fairness operations.}
}
\if{html}{\out{</div>}}
}
}
}
