% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compas.R
\name{mlr_tasks_compas}
\alias{mlr_tasks_compas}
\title{COMPAS Classification Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:TaskClassif]{mlr3::TaskClassif}.
}
\description{
Derived tasks: \cr
\itemize{
\item \code{compas} : A classification task for the \link{compas} data set with the protected attribute 'sex'.
}
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_tasks$get("compas")
tsk("compas")
}\if{html}{\out{</div>}}
}

