% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compas.R
\name{mlr_tasks_compas_race_binary}
\alias{mlr_tasks_compas_race_binary}
\title{COMPAS Classification Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:TaskClassif]{mlr3::TaskClassif}.
}
\description{
\itemize{
\item \code{compas_race_binary} : A classification task for the \link{compas} data set with the protected attribute 'race'.
The observations have been filtered, keeping only observations with race
\code{"Caucasian"} and \code{"African-American"}. The protected attribute has been set
to \code{"race"}.
}
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_tasks$get("compas_race_binary")
tsk("compas_race_binary")
}\if{html}{\out{</div>}}
}

