% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report_fairness}
\alias{report_fairness}
\title{Create a fairness report for a ML model}
\usage{
report_fairness(filename = "fairness_report.Rmd", objects, edit = FALSE)
}
\arguments{
\item{filename}{(\code{character})\cr
Filepath or name for new file that should be created.
Defaults to \code{"modelcard.Rmd"} which creates a new file called "modelcard.Rmd" in your local working directory.}

\item{objects}{(\code{list})\cr
A named list of objects required for the fairness report.
Objects are saved as \verb{<name>.RDS} in the new folder created for the report.
\itemize{
\item \code{task} :: The \code{\link{Task}} a report should be created for.
\item \code{resample_result} ::  A \link[mlr3:ResampleResult]{mlr3::ResampleResult} result to be analyzed.
\item \code{...} :: any other objects required for the report.
}}

\item{edit}{(\code{logical})\cr
\code{TRUE} to edit the template immediately.}
}
\description{
Instantiates a new R Markdown template with a skeleton of reported
metrics and visualizations that can be used to create a fairness report.
}
\examples{
\dontrun{
  task = tsk("compas")
  learner = lrn("classif.rpart", prob = TRUE)
  rr = resample(task, learner, rsmp("cv", folds = 5))
  report_fairness("documentation/fairness.Rmd", list(task = task, resampling_result = rr))
}
}
