% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_reports.R
\name{report_modelcard}
\alias{report_modelcard}
\title{Create a modelcard for documenting a ML model}
\usage{
report_modelcard(filename = "modelcard.Rmd", edit = TRUE)
}
\arguments{
\item{filename}{(\code{character})\cr
Filepath or name for new file that should be created.
Defaults to \code{"modelcard.Rmd"} which creates a new file called "modelcard.Rmd" in your local working directory.}

\item{edit}{(\code{logical})\cr
\code{TRUE} to edit the template immediately.}
}
\description{
Instantiates a new R Markdown template with a skeleton questionaire that allows for creating
dataset documentation for a given dataset.
The interplay of code and markdown text allowed throu R Markdown
allows for heavily improved documentation through the use of visualizations or summary statistics.
}
\details{
Uses the awesome markdown template created by Chris Garbin
\href{https://github.com/fau-masters-collected-works-cgarbin/model-card-template}{from Github}.
}
\examples{
\dontrun{
  report_modelcard("documentation/modelcard.Rmd")
}
}
\references{
Mitchell, Margaret, Wu, Simone, Zaldivar, Andrew, Barnes, Parker, Vasserman, Lucy, Hutchinson, Ben, Spitzer, Elena, Raji, Deborah I, Gebru, Timnit (2019).
\dQuote{Model cards for model reporting.}
In \emph{Proceedings of the conference on fairness, accountability, and transparency}, 220--229.
}
