% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_reports.R
\name{use_datasheet}
\alias{use_datasheet}
\title{Create a datasheet for documenting a dataset}
\usage{
use_datasheet(filename = "datasheet.Rmd", edit = TRUE)
}
\arguments{
\item{filename}{(\code{character})\cr
Filepath or name for new file that should be created.
Defaults to \code{"datasheet.Rmd"} which creates a new file called "datasheet.Rmd" in your local working directory.}

\item{edit}{(\code{logical})\cr
\code{TRUE} to edit the template immediately.}
}
\description{
Instantiates a new R Markdown template with a skeleton questionaire that allows for creating
a model card for a given ML model.
The interplay of code and markdown text allowed throu R Markdown
allows for heavily improved documentation through the use of visualizations or summary statistics.
Uses the awesome markdown template created by Chris Garbin
\href{https://github.com/fau-masters-collected-works-cgarbin/datasheet-for-dataset-template}{from Github}.
}
\examples{
\dontrun{
  use_datasheet"documentation/datasheet.Rmd")
}
}
\references{
Gebru, Timnit, Morgenstern, Jamie, Vecchione, Briana, Vaughan, Wortman J, Wallach, Hanna, III D, Hal, Crawford, Kate (2018).
\dQuote{Datasheets for datasets.}
\emph{arXiv preprint arXiv:1803.09010}.
}
