% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult.R
\docType{data}
\name{adult}
\alias{adult}
\alias{adult_test}
\alias{adult_train}
\title{Adult Dataset}
\source{
@misc{Dua:2019 ,
author = "Dua, Dheeru and Graff, Casey",
year = "2017",
title = "{UCI} Machine Learning Repository",
url = "http://archive.ics.uci.edu/ml",
institution = "University of California, Irvine, School of Information and Computer Sciences" }
}
\description{
Dataset used to predict whether income exceeds $50K/yr based on census data.
Also known as "Census Income" dataset
Train dataset contains 13 features and 30178 observations.
Test dataset contains 13 features and 15315 observations.
Target column is "target": A binary factor where 1: <=50K and 2: >50K for annual income.
The column \code{"sex"} is set as protected attribute.
}
\section{Pre-processing}{

\itemize{
\item \code{country} column has been removed since we expect them have no predictive power.
\item \code{fnlwgt} column has been removed since we expect them have no predictive power.
\item Rows containing \code{NA} in workclass and occupation have been removed.
}
}

\examples{
data("adult_test", package = "mlr3fairness")
data("adult_train", package = "mlr3fairness")
}
\keyword{data}
