% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{fairness_compare}
\alias{fairness_compare}
\title{Fairness Comparison}
\usage{
fairness_compare(predcitions, ...)
}
\arguments{
\item{predcitions}{TODO: fill it}
}
\description{
Fairness Comparison
}
\examples{
NULL
}
