% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{fairness_prediction_density}
\alias{fairness_prediction_density}
\title{Fairness Prediction Density Visualization}
\usage{
fairness_prediction_density(object, ...)
}
\arguments{
\item{object}{(\link{PredictionClassif})\cr (\link{ResampleResult})\cr (\link{BenchmarkResult})\cr
The binary class prediction object that will be evaluated. Only one data task is allowed for Resample Result and Benchmark Result.}

\item{...}{The arguments to be passed to methods, such as:
\itemize{
\item \code{task} (\link{TaskClassif})\cr
The data task that contains the protected column.
}}
}
\description{
This function specializes in visualizing the plots that will demonstrate the predicted probability density for each level in the protected column.
Which will potentially reveal the fairness problems in the dataset if the density differs significantly for each level.
The plot is a combination of boxplot and violin plot. So users could use the plot to detect both the outliers and distribution problems.
The y-axis shows the levels in protected columns. And the x-axis shows the predicted probability.
The title for the plot will demonstrate which class for predicted probability.
}
\examples{
library(mlr3fairness)
library(mlr3learners)
library(mlr3)
library(ggplot2)
library(data.table)

task = tsk("adult_train")
learner = lrn("classif.ranger", predict_type = "prob")
learner$train(task)
predictions = learner$predict(task)
fairness_prediction_density(predictions, task)
}
