% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fairness_tensor}
\alias{fairness_tensor}
\title{Compute the fairness tensor given a prediction and task.}
\usage{
fairness_tensor(object, task, ...)
}
\arguments{
\item{object}{\link[=`data.table`|`PredictionClassif`]{\code{data.table}|\code{PredictionClassif}}\cr
A data.table with columns \code{truth} and \code{prediction} or a \code{PredictionClassif}.}

\item{task}{\code{\link{TaskClassif}}\cr
A \code{\link{TaskClassif}}. Needs \code{col_role} \code{"pta"} to be set.}
}
\value{
\code{list} of confusion matrix for every group in \code{"pta"}.
}
\description{
A fairness tensor is a list-of-groupwise confusion matrices.
}
\examples{
library(mlr3)
t = tsk("compas")
prd = lrn("classif.rpart")$train(t)$predict(t)
fairness_tensor(prd, t)
}
