% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_operations.R
\name{groupdiff_tau}
\alias{groupdiff_tau}
\title{Groupwise Ratio}
\usage{
groupdiff_tau(x, y = NULL, FUN = "/")
}
\arguments{
\item{x}{\link{´numeric`} \cr
Measured performance in group 1.}

\item{y}{[´numeric\verb{] \\cr Measured performance in group 2. Defaults to }NULL`.}

\item{fun}{[character\verb{] \\cr Difference function, see }FUN\verb{in}outer()\verb{. Defaults to  }"\"`.}
}
\description{
Computes \code{min(x/y, y/x)} i.e. the smallest symmetric ratio between x and y
that is smaller than 1. If x is a vector, the symmetric ratio between all
elements in 'x' is computed.
}
