% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_operations.R
\name{groupdiff_tau}
\alias{groupdiff_tau}
\title{Groupwise Ratio}
\usage{
groupdiff_tau(x)
}
\arguments{
\item{x}{\code{numeric} \cr
Measured performance in group 1, 2, ...}
}
\description{
Computes \code{min(x/y, y/x)} i.e. the smallest symmetric ratio between x and y
that is smaller than 1. If x is a vector, the symmetric ratio between all
elements in 'x' is computed.
}
