% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupwise_abs_diff.R
\name{groupwise_abs_diff}
\alias{groupwise_abs_diff}
\title{Groupwise Absolute Differenence (Fairness Metrics)}
\usage{
groupwise_abs_diff(prediction, base_measure, data_task, ...)
}
\arguments{
\item{prediction}{The predictions of the learner.}

\item{base_measure}{The base measures used to evaluate.}

\item{data_task}{The data task for the fairness metric.}

\item{...}{}
}
\value{
Absolute difference of base measues between binary protected groups.
}
\description{
Groupwise Absolute Differenence (Fairness Metrics)
}
