% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupwise_diff.R
\name{groupwise_diff}
\alias{groupwise_diff}
\title{Groupwise Differenence (Fairness Metrics)}
\usage{
groupwise_diff(prediction, base_measure, data_task, ...)
}
\arguments{
\item{prediction}{The predictions of the learner.}

\item{base_measure}{The base measures used to evaluate.}

\item{data_task}{The data task for the fairness metric.}

\item{...}{Additional parameters}
}
\value{
Difference of base measues between binary protected groups.
}
\description{
Groupwise Differenence (Fairness Metrics)
}
