% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_fairness.R
\name{individual_fairness}
\alias{individual_fairness}
\title{Compute observation wise individual fairness.}
\usage{
individual_fairness(data, prediction, frac = 1)
}
\arguments{
\item{data}{\code{\link{data.table}} Variables to compare.}

\item{prediction}{\code{\link{Prediction}} Predictions to compare.}

\item{fraction}{\link{`numeric} For large datasets: Should computation be approximated by computing on only a fraction?}
}
\value{
A \code{\link{data.table}} with row_ids and the average individual fairness per observation.
}
\description{
The corresponding notion is "similar individuals should be treated similarly".
This is computed by comparing differences between predictions  and differences between features, using the
Gower distance.
}
