% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureFairness.R
\name{MeasureFairness}
\alias{MeasureFairness}
\title{Fairness Measures}
\description{
This measure specializes \code{\link[mlr3:Measure]{mlr3::Measure()}} to allow for measuring statistical group fairness:
A common approach to quantifying a model's fairness is to compute the difference between a
protected and an unprotected group according to some performance metric such as
\verb{classification error} (\link{mlr_measures_classif.ce}), \verb{false positive rate}
(\link{mlr_measures_classif.fpr}) or others.
This measure allows for the composition of a fairness-metric from such a "base_measure" by comparing
the group-wise scores. The operation for comparison (e.g., difference or quotient) can be
specified using the \code{operation} parameter.
}
\section{Overview}{

Composite measures encompasing multiple fairness metrics can be built using
\link{MeasureFairnessComposite}.

Predefined measures can be found in the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_measures}.
}

\examples{
# Create MeasureFairness to measure the Predictive Parity.
data_task = tsk("adult_train")
learner = lrn("classif.rpart", cp = .01)
learner$train(data_task)
measure = msr("fairness", base_measure = msr("classif.ppv"))
predictions = learner$predict(data_task)
predictions$score(measure, task = data_task)
}
\seealso{
groupdiff_tau groupdiff_absdiff
}
\section{Super class}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{MeasureFairness}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{base_measure}}{(\code{Measure()})\cr
The base measure to be used by the fairness measures,
e.g. \link{mlr_measures_classif.fpr} for the false positive rate.}

\item{\code{operation}}{(\verb{function()})\cr
The operation used to compute the difference. A function with args 'x' and 'y' that returns
a single value. Defaults to \code{abs(x - y)}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MeasureFairness$new()}}
\item \href{#method-clone}{\code{MeasureFairness$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate">}\href{../../mlr3/html/Measure.html#method-aggregate}{\code{mlr3::Measure$aggregate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format">}\href{../../mlr3/html/Measure.html#method-format}{\code{mlr3::Measure$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help">}\href{../../mlr3/html/Measure.html#method-help}{\code{mlr3::Measure$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print">}\href{../../mlr3/html/Measure.html#method-print}{\code{mlr3::Measure$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score">}\href{../../mlr3/html/Measure.html#method-score}{\code{mlr3::Measure$score()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureFairness$new(
  base_measure,
  operation = groupdiff_absdiff,
  minimize = TRUE,
  range = c(-Inf, Inf)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_measure}}{(\code{Measure()})\cr
The measure used to perform fairness operations.}

\item{\code{operation}}{(\code{function})\cr
The operation used to compute the difference. A function that returns
a single value given input: computed metric for each subgroup.
Defaults to \link{groupdiff_absdiff}.}

\item{\code{minimize}}{(\code{logical})\cr
Should the measure be minimized? Defaults to \code{TRUE}.}

\item{\code{range}}{(\code{numeric(2)})\cr
Range of the resulting measure. Defaults to \code{c(-Inf, Inf)}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureFairness$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
