% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3fairness-package}
\alias{mlr3fairness}
\alias{mlr3fairness-package}
\title{mlr3fairness: Fairness Auditing and Debiasing for mlr3}
\description{
Integrates fairness auditing and debiasing methods into the 'mlr3' ecosystem.
    This includes fairness metrics, visualizations and debiasing methods such as 'Reweighing' and 'Equalized Odds'.
    Integration with mlr3 allows for auditing of mlr3 models as well as convenient joint tuning of
    machine learning algorithms and debiasing methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3fairness.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3fairness}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3fairness/issues}
}

}
\author{
\strong{Maintainer}: Wei Siyi \email{weisiyi2@gmail.com}

Authors:
\itemize{
  \item Florian Pfisterer \email{pfistererf@googlemail.com} (\href{https://orcid.org/0000-0001-8867-762X}{ORCID})
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})
}

}
