% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3fairness-package}
\alias{mlr3fairness}
\alias{mlr3fairness-package}
\title{mlr3fairness: Fairness Auditing and Debiasing for 'mlr3'}
\description{
Integrates fairness auditing and bias mitigation methods for the 'mlr3' ecosystem. This includes fairness metrics, reporting tools, visualizations and bias mitigation techniques such as "Reweighing" described in 'Kamiran, Calders' (2012) \doi{10.1007/s10115-011-0463-8} and "Equalized Odds" described in 'Hardt et al.' (2016) \url{https://papers.nips.cc/paper/2016/file/9d2682367c3935defcb1f9e247a97c0d-Paper.pdf}. Integration with 'mlr3' allows for auditing of ML models as well as convenient joint tuning of machine learning algorithms and debiasing methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3fairness.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3fairness}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3fairness/issues}
}

}
\author{
\strong{Maintainer}: Florian Pfisterer \email{pfistererf@googlemail.com} (\href{https://orcid.org/0000-0001-8867-762X}{ORCID})

Authors:
\itemize{
  \item Wei Siyi \email{weisiyi2@gmail.com}
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})
}

}
\keyword{internal}
