% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureFairness.R
\docType{data}
\name{mlr_measures_fairness}
\alias{mlr_measures_fairness}
\alias{MeasureFairness}
\title{Fairness Measure}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 8 rows and 2 columns.
}
\usage{
mlr_measures_fairness
}
\description{
This measure specializes \code{\link[mlr3:Measure]{mlr3::Measure()}} to allow for measuring statistical group fairness:
A common approach to quantifying a model's fairness is to compute the difference between a protected and an
unprotected group according to some performance metric such as \verb{classification error},
\verb{false positive rate} or others.
This measure allows for the composition of a fairness-metric from such a \code{base_measure} by computing
group-wise difference. This difference metric can be specified using the  'operation' parameter.
}
\section{Overview}{

For an overview of measures see \code{\link{mlr_measures_fairness}}.
Composite measures encompasing multiple fairness metrics can be built using \code{\link{MeasureFairnessComposite}}.

Predefined measures can be found in the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_measures}.
See \code{\link{groupdiff_tau}} for a simple way to construct difference functions.
}

\section{Predefined measures}{

\code{mlr3fairness} comes with a set of predefined fairness measures listed below.
For full flexibility, \code{\link{MeasureFairness}} can be used to construct classical
group fairness measures based on a difference between a performance metrics across groups
by combining a performance measure with an operation for measuring differences.

\tabular{ll}{
fairness.EOd \tab Equalized Odds: Abs. difference between true positive and false positive rates across groups           \cr
fairness.fpr \tab Abs. difference in false positive rates across groups                                                  \cr
fairness.acc \tab Abs. difference in accuracy across groups (Overall accuracy equality)                               \cr
fairness.tpr \tab Abs. difference in True positive rates across groups                                                   \cr
fairness.ppv \tab Abs. difference in Positive predictive values across groups (Part of Conditional use accuracy equality)\cr
fairness.npv \tab Abs. difference in Negative predictive values across groups (Part of Conditional use accuracy equality)\cr
fairness.fp  \tab Abs. difference in False positives across groups (Part of Treatment equality)                          \cr
fairness.fn  \tab Abs. difference in False negatives across groups (Part of Treatment equality)
}
}

\examples{
  # Create MeasureFairness to measure the Predictive Parity.
  library(mlr3)
  data_task = tsk("adult_train")
  learner = lrn("classif.rpart", cp = .01)
  learner$train(data_task)
  measure = msr("fairness", base_measure = msr("classif.ppv"))
  predictions = learner$predict(data_task)
  predictions$score(measure, task = data_task)
# Predfined measures:
msr("fairness.EOd")
msr("fairness.fpr")
msr("fairness.acc")
msr("fairness.fnr")
msr("fairness.tpr")
msr("fairness.ppv")
msr("fairness.npv")
msr("fairness.fp")
msr("fairness.fn")
}
\seealso{
groupdiff_tau
}
\keyword{datasets}
\section{Super class}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{MeasureFairness}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{base_measure}}{(\code{Measure()})\cr
The base measure to be used by the fairness measures,
e.g. \link{mlr_measures_classif.fpr} for the false positive rate.}

\item{\code{operation}}{(\verb{function()})\cr
The operation used to compute the difference. A function with args 'x' and 'y' that returns
a single value. Defaults to \code{abs(x - y)}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MeasureFairness$new()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate">}\href{../../mlr3/html/Measure.html#method-aggregate}{\code{mlr3::Measure$aggregate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format">}\href{../../mlr3/html/Measure.html#method-format}{\code{mlr3::Measure$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help">}\href{../../mlr3/html/Measure.html#method-help}{\code{mlr3::Measure$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print">}\href{../../mlr3/html/Measure.html#method-print}{\code{mlr3::Measure$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score">}\href{../../mlr3/html/Measure.html#method-score}{\code{mlr3::Measure$score()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureFairness$new(
  base_measure,
  operation = function(x) {     max(abs(diff(x))) },
  minimize = TRUE,
  range = c(-Inf, Inf)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_measure}}{(\code{Measure()})\cr
The measure used to perform fairness operations.}

\item{\code{operation}}{(\code{function})\cr
The operation used to compute the difference. A function with args 'x' and 'y'(optional) that returns
a single value. Defaults to \code{max(abs(diff(x)))}.}

\item{\code{minimize}}{(\code{logical})\cr
Should the measure be minimized? Defaults to \code{TRUE}.}

\item{\code{range}}{(\code{numeric})\cr
Range of the resulting measure. Defaults to \code{c(-Inf, Inf)}.}
}
\if{html}{\out{</div>}}
}
}
}
