% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compas.R
\name{mlr_tasks_compas}
\alias{mlr_tasks_compas}
\title{COMPAS Classification Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskClassif}.
}
\description{
A classification task for the \link{compas} data set.
}
\section{Construction}{
\preformatted{mlr_tasks$get("compas")
tsk("compas")
}
}

