% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compas.R
\name{mlr_tasks_compas_race_binary}
\alias{mlr_tasks_compas_race_binary}
\title{COMPAS Classification Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskClassif}.
}
\description{
A classification task for the \link{compas} data set.
The observations have been filtered, keeping only observations with race
\code{"Caucasian"} and \code{"African-American"}. The protected attribute has been set
to \code{"race"}.
}
\section{Construction}{
\preformatted{mlr_tasks$get("compas_race_binary")
tsk("compas_race_binary")
}
}

